move pytorch and tensorflow files
